+++ src/Services/ResultPrinter/ResultPrinter.php
@@ -17,8 +17,8 @@ use webignition\BasilRunner\Model\TestOutput\Test as TestOutput;
 use webignition\BasilRunner\Services\ProjectRootPathProvider;
 use webignition\BasilRunner\Services\ResultPrinter\FailedAssertion\SummaryLineFactory;
 use webignition\BasilRunner\Services\ResultPrinter\FailedAssertion\SummaryHandler;
-use webignition\BasilRunner\Services\ResultPrinter\StepRenderer\StatementLineFactory;
 use webignition\BasilRunner\Services\ResultPrinter\StepRenderer\StepRenderer;
+use webignition\BasilRunner\Services\TestOutputRenderer\TestRenderer;
 
 class ResultPrinter extends Printer implements TestListener
 {
@@ -50,7 +50,7 @@ class ResultPrinter extends Printer implements TestListener
         $this->testRenderer = new TestRenderer($consoleOutputFactory);
         $this->stepRenderer = new StepRenderer(
             $consoleOutputFactory,
-            new StatementLineFactory($consoleOutputFactory),
+            new StatementLineRenderer($consoleOutputFactory),
             new SummaryHandler(
                 DomIdentifierFactory::createFactory(),
                 new SummaryLineFactory($consoleOutputFactory)
